enh(Serializer): split from Markdown extension
We are now using this for serializing plaintext as well.
