+++ src/extensions/PlainText.js
@@ -22,9 +22,10 @@
 
 import { Extension } from '@tiptap/core'
 
+import PlainTextDocument from './../nodes/PlainTextDocument.js'
+import Serializer from './../extensions/Serializer.js'
 /* eslint-disable import/no-named-as-default */
 import Text from '@tiptap/extension-text'
-import PlainTextDocument from './../nodes/PlainTextDocument.js'
 
 export default Extension.create({
 	name: 'PlainText',
@@ -32,6 +33,7 @@ export default Extension.create({
 	addExtensions() {
 		return [
 			PlainTextDocument,
+			Serializer,
 			Text,
 		]
 	},
